Build up basic CRUD for User
In order to do data manipulation for User in database,
we need to design a proper structure for database
operations. We try to simulate structure like MVC.

Model is the abstraction for tables in database, and
Service is the interface for operations on table,
at the same time we have ServiceImpl as the
implementation for Service.

Controller is where we receive request and perform
the corresponding operation to database via Service
. Last but not least, Router acts as the abstraction
 of routes for all the backend API.

In this commit, we implement the basics CRUD method
with the structure mentioned above.